feat: create user role model and repository
The idea behind those two is to separate the permission check from the
JWT, due to security issues. The permission check will be done querying
the userrole in database to check if the user truly have that level of
authorization. However, to avoid resource overhead, the checking only
will be made if the user say that he got high level permission such as
admininistrator